Move and split tests to subfolder
